Error: Replace this use of System.out or System.err by a logger.
In order to overcome this error we have used java.util.Logger class
utility methods to log the various levels of information.

Why to change? since developers use System.out or System.err to debug
the information to console in development environment. But console
messages are of no use in production environment. So it is necessary to
replace all System.out or System.err  statements to properly configured
Loggers before deploying to production servers.