Feature : Add connection checker to external recommender settings #726
Solution : Instead of implementing a separate button for checking the
external connection we have used the existing save button to verify the
form validation. We have developed a custom form validator to remote URL
which takes URL as input and establishes connection and extracts the
headers. From the headers we receive a status code which when equals to
HTTP_OK(200), the connection is successful. If the connection is not
successfully established, the form won't be saved and returns an error
message with the appropriate status code.